requirements.txt
slack-sdk==3.26.1
python-dotenv==1.0.0
pyinotify==0.9.6 